Update move_forward and added MAX_SPEED
For move_forward of Snake, NormalGhost and BigGhost
Their number of unit moved should still be 1 instead of their corresponding speed. In order for them to move "faster", we should do something on the game engine. For example, we should use update their location using game tick

#include <cmath>
// Assume that is a signal (int) being emitted every game tick
public slot:
    void move_forward_slot() {
    if (gameTime % static_cast<int>(1 / speed * 10)) {
        move_forward()
        }
    }

So the maximum speed for all moving entity will be 10 in this case.